+++ src/chunk_index.h
@@ -38,8 +38,6 @@ extern void ts_chunk_index_create_from_adjusted_index_info(int32 hypertable_id,
 														   IndexInfo *indexinfo);
 extern TSDLLEXPORT void ts_chunk_index_create_all(int32 hypertable_id, Oid hypertable_relid,
 												  int32 chunk_id, Oid chunkrelid);
-extern Oid ts_chunk_index_create_from_stmt(IndexStmt *stmt, int32 chunk_id, Oid chunkrelid,
-										   int32 hypertable_id, Oid hypertable_indexrelid);
 extern int ts_chunk_index_delete(int32 chunk_id, const char *indexname, bool drop_index);
 extern int ts_chunk_index_delete_by_chunk_id(int32 chunk_id, bool drop_index);
 extern void ts_chunk_index_delete_by_name(const char *schema, const char *index_name,